Broke treetools into component submodules
